Split further sharedClassInfo.hpp to dumpTimeSharedClassInfo.hpp and runTimeSharedClassInfo.hpp to conform to simple rule of separation of definition and implementation
