+++ make/hotspot/lib/JvmFeatures.gmk
@@ -119,7 +119,9 @@ ifneq ($(call check-jvm-feature, cds), true)
       cdsProtectionDomain.cpp \
       classLoaderDataShared.cpp \
       classLoaderExt.cpp \
-      sharedClassInfo.cpp \
+      dumpTimeSharedClassInfo.cpp \
+      lambdaProxyClassDictionary.cpp \
+      runTimeSharedClassInfo.cpp \
       systemDictionaryShared.cpp
   JVM_EXCLUDE_PATTERNS += cds/
 endif
