+++ src/hotspot/share/classfile/systemDictionaryShared.cpp
@@ -29,10 +29,8 @@
 #include "cds/dynamicArchive.hpp"
 #include "cds/filemap.hpp"
 #include "cds/heapShared.hpp"
-#include "cds/lambdaProxyClassInfo.hpp"
 #include "cds/cdsProtectionDomain.hpp"
 #include "cds/metaspaceShared.hpp"
-#include "cds/sharedClassInfo.hpp"
 #include "classfile/classFileStream.hpp"
 #include "classfile/classLoader.hpp"
 #include "classfile/classLoaderData.inline.hpp"
